@@ -9,7 +9,7 @@ AboutPage.configure do |config|
   config.fedora           = AboutPage::Fedora.new(ActiveFedora.fedora.connection)
   config.solr             = AboutPage::Solr.new(ActiveFedora.solr.conn, :numDocs => 1)
   config.database         = Avalon::About::Database.new(User)
-  config.mediainfo        = Avalon::About::MediaInfo.new(:version => '>=0.7.59')
+  config.mediainfo        = Settings&.mediainfo&.path ? Avalon::About::MediaInfo.new(path: Settings.mediainfo.path) : Avalon::About::MediaInfo.new()
   config.streaming_server = Avalon::About::HLSServer.new(Settings.streaming.http_base)
   config.sidekiq          = Avalon::About::Sidekiq.new(numProcesses: 1)
   config.redis            = Avalon::About::Redis.new(Redis.new(Rails.application.config.cache_store[1]))
