@@ -9,7 +9,7 @@ require 'resolv-replace'
 Bundler.require(*Rails.groups)
 
 module Avalon
-  VERSION = '7.8.0'
+  VERSION = '8.0.0'
 
   class Application < Rails::Application
     require 'avalon/configuration'
@@ -19,11 +19,16 @@ module Avalon
     end
 
     # Initialize configuration defaults for originally generated Rails version.
-    config.load_defaults 6.0
+    config.load_defaults 7.2
 
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration should go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded.
+
+    # Please, add to the `ignore` list any other `lib` subdirectories that do
+    # not contain `.rb` files, or that should not be reloaded or eager loaded.
+    # Common ones are `templates`, `generators`, or `middleware`, for example.
+    # config.autoload_lib(ignore: %w[assets avalon capistrano tasks])
 
     # Set Time.zone default to the specified zone and make Active Record auto-convert to this zone.
     # Run "rake -D time" for a list of tasks for finding time zone names. Default is UTC.
@@ -39,21 +44,31 @@ module Avalon
 
     config.action_dispatch.default_headers = { 'X-Frame-Options' => 'ALLOWALL' }
 
+    # We have a number of serializers in place that have not previously had a :coder defined.
+    # Setting our global default to the old default :coder should maintain compatibility.
+    config.active_record.default_column_serializer = YAML
+
+    # Rails recommends having this set to false, especially in zeitwerk mode. However, that
+    # currently causes issues with the Samvera gems (hydra-head, Blacklight)
+    config.add_autoload_paths_to_load_path = true
+
     config.middleware.insert_before 0, Rack::Cors do
       allow do
         origins { |source| true }
         resource '/avalon_marker/*', headers: :any, credentials: true, methods: [:get, :post, :put, :delete]
         resource '/media_objects/*/manifest*', headers: :any, methods: [:get]
         resource '/master_files/*/thumbnail', headers: :any, methods: [:get]
-        resource '/master_files/*/transcript/*/*', headers: :any, methods: [:get]
+        resource '/master_files/*/transcript/*', headers: :any, methods: [:get]
         resource '/master_files/*/structure.json', headers: :any, methods: [:get, :post, :delete]
         resource '/master_files/*/waveform.json', headers: :any, methods: [:get]
         resource '/master_files/*/*.m3u8', headers: :any, credentials: true, methods: [:get, :head]
         resource '/master_files/*/captions', headers: :any, methods: [:get]
         resource '/master_files/*/supplemental_files/*', headers: :any, methods: [:get]
-        resource '/playlists/*/manifest.json', headers: :any, credentials: true, methods: [:get]
-        resource '/timelines/*/manifest.json', headers: :any, methods: [:get, :post]
+        resource '/playlists/*/manifest*', headers: :any, credentials: true, methods: [:get]
+        resource '/timelines/*/manifest*', headers: :any, methods: [:get, :post]
         resource '/master_files/*/search', headers: :any, methods: [:get]
+        resource '/rails/active_storage/blobs/*/*/*', headers: :any, methods: [:get]
+        resource '/rails/active_storage/disk/*/*', headers: :any, methods: [:get]
       end
     end
 
