@@ -26,7 +26,8 @@ FactoryBot.define do
     video_bitrate { '4000000.0' }
     audio_codec { 'AAC' }
     audio_bitrate { '163842.0' }
-    absolute_location { 'file:///srv/avalon/content/6f69c008-06a4-4bad-bb60-26297f0b4c06/35bddaa0-fbb4-404f-ab76-58f22921529c/warning.mp4' }
+    mime_type { nil }
+    derivativeFile { 'file:///srv/avalon/content/6f69c008-06a4-4bad-bb60-26297f0b4c06/35bddaa0-fbb4-404f-ab76-58f22921529c/warning.mp4' }
 
     trait :with_master_file do
       after(:create) do |d|
