@@ -24,12 +24,18 @@ describe ApplicationJob do
     end
 
     context 'raise_on_connection_error disabled' do
-      let(:request_context) { { body: "request_context" } }
+      let(:request_context) { { uri: Addressable::URI.parse("http://example.com"), body: "request_context" } }
       let(:e) { { body: "error_response" } }
 
       [RSolr::Error::ConnectionRefused, RSolr::Error::Timeout, Blacklight::Exceptions::ECONNREFUSED, Faraday::ConnectionFailed].each do |error_code|
         it "rescues #{error_code} errors" do
-          raised_error = error_code == RSolr::Error::Timeout ? error_code.new(request_context, e) : error_code 
+          raised_error = if error_code == RSolr::Error::ConnectionRefused
+                           error_code.new(request_context)
+                         elsif error_code == RSolr::Error::Timeout
+                           error_code.new(request_context, e)
+                         else
+                           error_code
+                         end
           allow(Settings.app_job.solr_and_fedora).to receive(:raise_on_connection_error).and_return(false)
           allow_any_instance_of(described_class).to receive(:perform).and_raise(raised_error)
           allow_any_instance_of(Exception).to receive(:backtrace).and_return(["Test trace"])
@@ -41,12 +47,18 @@ describe ApplicationJob do
     end
 
     context 'raise_on_connection_error enabled' do
-      let(:request_context) { { body: "request_context" } }
+      let(:request_context) { { uri: Addressable::URI.parse("http://example.com"), body: "request_context" } }
       let(:e) { { body: "error_response" } }
 
       [RSolr::Error::ConnectionRefused, RSolr::Error::Timeout, Blacklight::Exceptions::ECONNREFUSED, Faraday::ConnectionFailed].each do |error_code|
         it "raises #{error_code} errors" do
-          raised_error = error_code == RSolr::Error::Timeout ? error_code.new(request_context, e) : error_code 
+          raised_error = if error_code == RSolr::Error::ConnectionRefused
+                           error_code.new(request_context)
+                         elsif error_code == RSolr::Error::Timeout
+                           error_code.new(request_context, e)
+                         else
+                           error_code
+                         end 
           allow(Settings.app_job.solr_and_fedora).to receive(:raise_on_connection_error).and_return(true)
           allow_any_instance_of(described_class).to receive(:perform).and_raise(raised_error)
           allow_any_instance_of(Exception).to receive(:backtrace).and_return(["Test trace"])
